+++ crates/icp-cli/src/commands/build.rs
@@ -4,7 +4,10 @@ use icp::context::{Context, EnvironmentSelection};
 
 use tracing::info;
 
-use crate::{operations::build::build_many_with_progress_bar, options::EnvironmentOpt};
+use crate::{
+    operations::build::build_many_with_progress_bar,
+    options::{EnvironmentOpt, arg_struct_change_help},
+};
 
 /// Build canisters
 #[derive(Debug, Args)]
@@ -13,12 +16,18 @@ pub(crate) struct BuildArgs {
     pub(crate) canisters: Vec<String>,
 
     #[command(flatten)]
-    pub(crate) environment: EnvironmentOpt,
+    pub(crate) environment: BuildEnvironmentOpt,
 }
+
+arg_struct_change_help!(
+    EnvironmentOpt => BuildEnvironmentOpt,
+    arg = "environment",
+    help = "Override the environment to build for. By default, the local environment is used."
+);
 
 pub(crate) async fn exec(ctx: &Context, args: &BuildArgs) -> Result<(), anyhow::Error> {
     // Get environment selection
-    let environment_selection: EnvironmentSelection = args.environment.clone().into();
+    let environment_selection: EnvironmentSelection = args.environment.0.clone().into();
 
     // Load target environment
     let env = ctx.get_environment(&environment_selection).await?;
@@ -48,6 +57,7 @@ pub(crate) async fn exec(ctx: &Context, args: &BuildArgs) -> Result<(), anyhow::
 
     build_many_with_progress_bar(
         canisters_to_build,
+        environment_selection.name(),
         ctx.builder.clone(),
         ctx.artifacts.clone(),
         &ctx.dirs.package_cache()?,